+++ src/config.py
@@ -8,8 +8,9 @@ class Config:
 
     def load(self):
         """Loads the config file into memory."""
-        path = os.path.dirname(os.path.realpath(__file__)).replace("\\", "/")
-        with open(f"{path}/config.cfg", "r", encoding="utf-8") as f:
+        path = os.path.dirname(os.path.dirname(os.path.realpath(__file__))).replace("\\", "/")
+        config_path = os.path.join(path, "config/config.cfg")
+        with open(config_path, "r", encoding="utf-8") as f:
             self.config = json.load(f)
 
     def get(self, key):
@@ -21,5 +22,4 @@ class Config:
         Returns:
           str: the value of the key
         """
-        self.load()
         return self.config.get(key)