+++ src/tasks/Math.py
@@ -1,5 +1,6 @@
 import math
 import string
+from typing import List
 
 
 def solve_expr(expr: str):